test: add Inventory module unit tests (17/55 passing) - fix DTOs and service impl

+++ src/main/java/com/project/inventory/service/impl/InventoryServiceImpl.java
@@ -15,6 +15,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -232,7 +233,7 @@ public class InventoryServiceImpl implements InventoryService {
                 .id(inventory.getId())
                 .productId(inventory.getProduct().getId())
                 .productName(inventory.getProduct().getName())
-                .productPrice(inventory.getProduct().getPrice())
+                .productPrice(BigDecimal.valueOf(inventory.getProduct().getPrice()))
                 .stock(inventory.getStock())
                 .reserved(inventory.getReserved())
                 .available(inventory.getAvailable())
